2018_05_07
서버와 클라이언트 불렛 생성 및 이동 버그수정.

+++ 최규화/Mod_Server/MT_EXAM01/Project04/BulletObject.cpp
@@ -1,7 +1,7 @@
 #include "BulletObject.h"
 
 BulletObject::BulletObject(const unsigned short& master_id, const unsigned short& my_id,
-						   const Position& pos, const Rotation& rot, float bulltime,
+						   const Position& pos, const Rotation& rot, double bulltime,
 						   Vel3f& vel, const unsigned char& type, const Position3D& endpt)
 {
 	pe = new PhysicalEffect();
